draupadi-on-dais/Vaani.ai/requirements.txt
crewai
groq
pydantic
crewai-tools
langgraph
langchain-community
langchain-groq
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
selenium
requests